# [PERF] project_timesheet_holidays: Speed up `_search_is_timeoff_task`