Description
-----------
Fixed `_search_is_timeoff_task` to properly handle the `value` parameter which previously ignored its value and always searched for timeoff tasks.

Improved performance by:
- Replacing `_read_group` with `search + DISTINCT` to avoid unnecessary ordering.
- Adding a partial index on timeoff timesheets to optimize lookups without scanning holiday/leave indexes. The index stays small due to low timeoff vs regular timesheet ratio.

Benchmark
---------
On odoo.com, the `web_name_search` of tasks from the timesheet timer on a consulting project took:

| Before | After  | Speed up |
|--------|--------|----------|
| 1s     | ~120ms | 8x       |

Reference
---------
task-4801619

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr